+++ packages/koni_jxl/lib/koni_jxl.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'src/decoder.dart' show JxlDecoder;
+export 'src/encode/encoder.dart' show JxlEncoder;
 export 'src/exceptions.dart';
 export 'src/jxl_image.dart' show JxlAnimation, JxlImage;
 export 'src/jxl_info.dart' show JxlInfo;